+++ src/main/java/io/brachu/johann/cli/CliRunner.java
@@ -1,11 +1,14 @@
 package io.brachu.johann.cli;
 
+import java.io.File;
 import java.io.IOException;
+import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
+import com.google.common.collect.ImmutableMap;
 import io.brachu.johann.cli.exception.ExecutionTimedOutException;
 import io.brachu.johann.cli.exception.NonZeroExitCodeException;
 import org.slf4j.Logger;
@@ -15,14 +18,40 @@ final class CliRunner {
 
     private static final Logger log = LoggerFactory.getLogger(CliRunner.class);
 
-    private CliRunner() {
+    private final String[] cmd;
+    private Map<String, String> env;
+    private File workDir;
+    private Consumer<CliProcess> onProcessStart;
+    private boolean verbose;
+
+    CliRunner(String[] cmd) {
+        this.cmd = cmd;
+        env = Collections.emptyMap();
     }
 
-    static String exec(String[] cmd, Map<String, String> env, Consumer<CliProcess> onProcessStart, boolean verbose)
-            throws IOException, InterruptedException, NonZeroExitCodeException, ExecutionTimedOutException {
+    CliRunner env(Map<String, String> env) {
+        CliRunner.this.env = ImmutableMap.copyOf(env);
+        return this;
+    }
 
+    CliRunner workDir(File workDir) {
+        this.workDir = workDir;
+        return this;
+    }
+
+    CliRunner onProcessStart(Consumer<CliProcess> onProcessStart) {
+        CliRunner.this.onProcessStart = onProcessStart;
+        return this;
+    }
+
+    CliRunner verbose(boolean verbose) {
+        CliRunner.this.verbose = verbose;
+        return this;
+    }
+
+    String exec() throws IOException, InterruptedException, NonZeroExitCodeException, ExecutionTimedOutException {
         log(cmd, env);
-        CliProcess process = startProcess(cmd, env, verbose);
+        CliProcess process = startProcess();
         onProcessStart.accept(process);
 
         if (process.waitFor(5, TimeUnit.MINUTES)) {
@@ -39,16 +68,19 @@ final class CliRunner {
         }
     }
 
-    private static CliProcess startProcess(String[] cmd, Map<String, String> env, boolean verbose) throws IOException {
+    private CliProcess startProcess() throws IOException {
         ProcessBuilder bp = new ProcessBuilder(cmd);
+        bp.directory(workDir);
         bp.environment().putAll(env);
         return new CliProcess(bp.start(), verbose);
     }
 
-    private static void log(String[] cmd, Map<String, String> env) {
+    private void log(String[] cmd, Map<String, String> env) {
         if (log.isTraceEnabled()) {
             log.trace("Running command: {}", String.join(" ", cmd));
-            log.trace("Env variables: {}", env.entrySet().stream().map(entry -> entry.getKey() + "=" + entry.getValue()).collect(Collectors.joining(" ")));
+            log.trace("Env variables: {}", env.entrySet().stream()
+                    .map(entry -> entry.getKey() + "=" + entry.getValue())
+                    .collect(Collectors.joining(" ")));
         }
     }